+++ src/lib.rs
@@ -17,6 +17,7 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 #[macro_use]
+#[allow(unused_imports)]
 extern crate alloc;
 
 pub mod nodes;
@@ -30,13 +31,8 @@ pub mod proof;
 mod mask;
 pub use mask::TrieMask;
 
-#[cfg(feature = "std")]
-use hashbrown as _;
-#[cfg(feature = "std")]
-pub use std::collections::HashMap;
-
-#[cfg(not(feature = "std"))]
-pub use hashbrown::HashMap;
+#[doc(hidden)]
+pub use alloy_primitives::map::HashMap;
 
 #[doc(no_inline)]
 pub use nybbles::{self, Nibbles};