+++ src/nodes/mod.rs
@@ -1,7 +1,7 @@
 //! Various branch nodes produced by the hash builder.
 
-use alloy_primitives::{keccak256, Bytes, B256};
-use alloy_rlp::{length_of_length, Buf, Decodable, Encodable, Header, EMPTY_STRING_CODE};
+use alloy_primitives::B256;
+use alloy_rlp::{Decodable, Encodable, Header, EMPTY_STRING_CODE};
 use core::ops::Range;
 use nybbles::Nibbles;
 use smallvec::SmallVec;
@@ -18,12 +18,17 @@ pub use extension::{ExtensionNode, ExtensionNodeRef};
 mod leaf;
 pub use leaf::{LeafNode, LeafNodeRef};
 
+mod rlp;
+pub use rlp::RlpNode;
+
 /// The range of valid child indexes.
 pub const CHILD_INDEX_RANGE: Range<u8> = 0..16;
 
 /// Enum representing an MPT trie node.
 #[derive(PartialEq, Eq, Clone, Debug)]
 pub enum TrieNode {
+    /// Variant representing empty root node.
+    EmptyRoot,
     /// Variant representing a [BranchNode].
     Branch(BranchNode),
     /// Variant representing a [ExtensionNode].
@@ -33,16 +38,22 @@ pub enum TrieNode {
 }
 
 impl Encodable for TrieNode {
+    #[inline]
     fn encode(&self, out: &mut dyn alloy_rlp::BufMut) {
         match self {
+            Self::EmptyRoot => {
+                out.put_u8(EMPTY_STRING_CODE);
+            }
             Self::Branch(branch) => branch.encode(out),
             Self::Extension(extension) => extension.encode(out),
             Self::Leaf(leaf) => leaf.encode(out),
         }
     }
 
+    #[inline]
     fn length(&self) -> usize {
         match self {
+            Self::EmptyRoot => 1,
             Self::Branch(branch) => branch.length(),
             Self::Extension(extension) => extension.length(),
             Self::Leaf(leaf) => leaf.length(),
@@ -52,101 +63,91 @@ impl Encodable for TrieNode {
 
 impl Decodable for TrieNode {
     fn decode(buf: &mut &[u8]) -> alloy_rlp::Result<Self> {
-        let mut bytes = Header::decode_bytes(buf, true)?;
-
-        let mut items = vec![];
-        while !bytes.is_empty() {
-            // Decode header without advancing.
-            let Header { payload_length, .. } = Header::decode(&mut &bytes[..])?;
-            // This is a workaround for a footgun in header representation. If the payload
-            // length is 1, then `length_of_length` would also be 1 leading to total
-            // length of 2 which is incorrect for RLP value of 1 byte.
-            let len = if payload_length == 1 {
-                1 // If payload length is 1 byte, then there is no header
-            } else {
-                payload_length + length_of_length(payload_length)
-            };
-            items.push(bytes[..len].to_vec());
-            bytes.advance(len);
-        }
+        let mut items = match Header::decode_raw(buf)? {
+            alloy_rlp::PayloadView::List(list) => list,
+            alloy_rlp::PayloadView::String(val) => {
+                return if val.is_empty() {
+                    Ok(Self::EmptyRoot)
+                } else {
+                    Err(alloy_rlp::Error::UnexpectedString)
+                }
+            }
+        };
 
         // A valid number of trie node items is either 17 (branch node)
         // or 2 (extension or leaf node).
-        if items.len() == 17 {
-            let mut branch = BranchNode::default();
-            for (idx, item) in items.into_iter().enumerate() {
-                if idx == 16 {
-                    if item != [EMPTY_STRING_CODE] {
-                        return Err(alloy_rlp::Error::Custom(
-                            "branch node values are not supported",
-                        ));
+        match items.len() {
+            17 => {
+                let mut branch = BranchNode::default();
+                for (idx, item) in items.into_iter().enumerate() {
+                    if idx == 16 {
+                        if item != [EMPTY_STRING_CODE] {
+                            return Err(alloy_rlp::Error::Custom(
+                                "branch node values are not supported",
+                            ));
+                        }
+                    } else if item != [EMPTY_STRING_CODE] {
+                        branch.stack.push(RlpNode::from_raw_rlp(item)?);
+                        branch.state_mask.set_bit(idx as u8);
                     }
-                } else if item != [EMPTY_STRING_CODE] {
-                    branch.stack.push(item);
-                    branch.state_mask.set_bit(idx as u8);
                 }
+                Ok(Self::Branch(branch))
             }
-            return Ok(Self::Branch(branch));
-        }
+            2 => {
+                let mut key = items.remove(0);
 
-        if items.len() == 2 {
-            let key = items.remove(0);
+                let encoded_key = Header::decode_bytes(&mut key, false)?;
+                if encoded_key.is_empty() {
+                    return Err(alloy_rlp::Error::Custom("trie node key empty"));
+                }
 
-            let encoded_key = Bytes::decode(&mut &key[..])?;
-            if encoded_key.is_empty() {
-                return Err(alloy_rlp::Error::Custom("trie node key empty"));
+                // extract the high order part of the nibble to then pick the odd nibble out
+                let key_flag = encoded_key[0] & 0xf0;
+                // Retrieve first byte. If it's [Some], then the nibbles are odd.
+                let first = match key_flag {
+                    ExtensionNode::ODD_FLAG | LeafNode::ODD_FLAG => Some(encoded_key[0] & 0x0f),
+                    ExtensionNode::EVEN_FLAG | LeafNode::EVEN_FLAG => None,
+                    _ => return Err(alloy_rlp::Error::Custom("node is not extension or leaf")),
+                };
+
+                let key = unpack_path_to_nibbles(first, &encoded_key[1..]);
+                let node = if key_flag == LeafNode::EVEN_FLAG || key_flag == LeafNode::ODD_FLAG {
+                    Self::Leaf(LeafNode::new(key, RlpNode::decode(&mut items.remove(0))?))
+                } else {
+                    // We don't decode value because it is expected to be RLP encoded.
+                    Self::Extension(ExtensionNode::new(
+                        key,
+                        RlpNode::from_raw_rlp(items.remove(0))?,
+                    ))
+                };
+                Ok(node)
             }
-
-            // extract the high order part of the nibble to then pick the odd nibble out
-            let key_flag = encoded_key[0] & 0xf0;
-            // Retrieve first byte. If it's [Some], then the nibbles are odd.
-            let first = match key_flag {
-                ExtensionNode::ODD_FLAG | LeafNode::ODD_FLAG => Some(encoded_key[0] & 0x0f),
-                ExtensionNode::EVEN_FLAG | LeafNode::EVEN_FLAG => None,
-                _ => return Err(alloy_rlp::Error::Custom("node is not extension or leaf")),
-            };
-
-            let key = unpack_path_to_nibbles(first, &encoded_key[1..]);
-            let node = if key_flag == LeafNode::EVEN_FLAG || key_flag == LeafNode::ODD_FLAG {
-                Self::Leaf(LeafNode::new(key, Bytes::decode(&mut &items.remove(0)[..])?.to_vec()))
-            } else {
-                // We don't decode value because it is expected to be RLP encoded.
-                Self::Extension(ExtensionNode::new(key, items.remove(0)))
-            };
-            return Ok(node);
+            _ => Err(alloy_rlp::Error::Custom("invalid number of items in the list")),
         }
-
-        Err(alloy_rlp::Error::Custom("invalid number of items in the list"))
     }
 }
 
 impl TrieNode {
-    /// RLP encodes the node and returns either RLP(Node) or RLP(keccak(RLP(node))).
-    pub fn rlp(&self, buf: &mut Vec<u8>) -> Vec<u8> {
-        self.encode(buf);
-        rlp_node(buf)
+    /// RLP-encodes the node and returns either `rlp(node)` or `rlp(keccak(rlp(node)))`.
+    #[inline]
+    pub fn rlp(&self, rlp: &mut Vec<u8>) -> RlpNode {
+        self.encode(rlp);
+        RlpNode::from_rlp(rlp)
     }
 }
 
-/// Given an RLP encoded node, returns either self as RLP(node) or RLP(keccak(RLP(node)))
+/// Given an RLP-encoded node, returns it either as `rlp(node)` or `rlp(keccak(rlp(node)))`.
 #[inline]
-pub(crate) fn rlp_node(rlp: &[u8]) -> Vec<u8> {
-    if rlp.len() < B256::len_bytes() {
-        rlp.to_vec()
-    } else {
-        word_rlp(&keccak256(rlp))
-    }
+#[deprecated = "use `RlpNode::from_rlp` instead"]
+pub fn rlp_node(rlp: &[u8]) -> RlpNode {
+    RlpNode::from_rlp(rlp)
 }
 
-/// Optimization for quick encoding of a 32-byte word as RLP.
-// TODO: this could return [u8; 33] but Vec is needed everywhere this function is used
+/// Optimization for quick RLP-encoding of a 32-byte word.
 #[inline]
-pub fn word_rlp(word: &B256) -> Vec<u8> {
-    // Gets optimized to alloc + write directly into it: https://godbolt.org/z/rfWGG6ebq
-    let mut arr = [0; 33];
-    arr[0] = EMPTY_STRING_CODE + 32;
-    arr[1..].copy_from_slice(word.as_slice());
-    arr.to_vec()
+#[deprecated = "use `RlpNode::word_rlp` instead"]
+pub fn word_rlp(word: &B256) -> RlpNode {
+    RlpNode::word_rlp(word)
 }
 
 /// Unpack node path to nibbles.
@@ -174,9 +175,10 @@ pub(crate) fn unpack_path_to_nibbles(first: Option<u8>, rest: &[u8]) -> Nibbles 
     Nibbles::from_vec_unchecked(nibbles)
 }
 
-/// Encodes a given path leaf as a compact array of bytes, where each byte represents two
-/// "nibbles" (half-bytes or 4 bits) of the original hex data, along with additional information
-/// about the leaf itself.
+/// Encodes a given path leaf as a compact array of bytes.
+///
+/// In resulted array, each byte represents two "nibbles" (half-bytes or 4 bits) of the original hex
+/// data, along with additional information about the leaf itself.
 ///
 /// The method takes the following input:
 /// `is_leaf`: A boolean value indicating whether the current node is a leaf node or not.
@@ -257,14 +259,33 @@ mod tests {
     use alloy_primitives::hex;
 
     #[test]
+    fn rlp_empty_root_node() {
+        let empty_root = TrieNode::EmptyRoot;
+        let rlp = empty_root.rlp(&mut vec![]);
+        assert_eq!(rlp[..], hex!("80"));
+        assert_eq!(TrieNode::decode(&mut &rlp[..]).unwrap(), empty_root);
+    }
+
+    #[test]
+    fn rlp_zero_value_leaf_roundtrip() {
+        let leaf = TrieNode::Leaf(LeafNode::new(
+            Nibbles::from_nibbles_unchecked(hex!("0604060f")),
+            RlpNode::from_raw(&alloy_rlp::encode(alloy_primitives::U256::ZERO)).unwrap(),
+        ));
+        let rlp = leaf.rlp(&mut vec![]);
+        assert_eq!(rlp[..], hex!("c68320646f8180"));
+        assert_eq!(TrieNode::decode(&mut &rlp[..]).unwrap(), leaf);
+    }
+
+    #[test]
     fn rlp_trie_node_roundtrip() {
         // leaf
         let leaf = TrieNode::Leaf(LeafNode::new(
             Nibbles::from_nibbles_unchecked(hex!("0604060f")),
-            hex!("76657262").to_vec(),
+            RlpNode::from_raw(&hex!("76657262")).unwrap(),
         ));
         let rlp = leaf.rlp(&mut vec![]);
-        assert_eq!(rlp, hex!("c98320646f8476657262"));
+        assert_eq!(rlp[..], hex!("c98320646f8476657262"));
         assert_eq!(TrieNode::decode(&mut &rlp[..]).unwrap(), leaf);
 
         // extension
@@ -272,21 +293,21 @@ mod tests {
         hex!("76657262").to_vec().as_slice().encode(&mut child);
         let extension = TrieNode::Extension(ExtensionNode::new(
             Nibbles::from_nibbles_unchecked(hex!("0604060f")),
-            child,
+            RlpNode::from_raw(&child).unwrap(),
         ));
         let rlp = extension.rlp(&mut vec![]);
-        assert_eq!(rlp, hex!("c98300646f8476657262"));
+        assert_eq!(rlp[..], hex!("c98300646f8476657262"));
         assert_eq!(TrieNode::decode(&mut &rlp[..]).unwrap(), extension);
 
         // branch
         let branch = TrieNode::Branch(BranchNode::new(
-            core::iter::repeat(word_rlp(&B256::repeat_byte(23))).take(16).collect(),
+            core::iter::repeat(RlpNode::word_rlp(&B256::repeat_byte(23))).take(16).collect(),
             TrieMask::new(u16::MAX),
         ));
         let mut rlp = vec![];
         let rlp_node = branch.rlp(&mut rlp);
         assert_eq!(
-            rlp_node,
+            rlp_node[..],
             hex!("a0bed74980bbe29d9c4439c10e9c451e29b306fe74bcf9795ecf0ebbd92a220513")
         );
         assert_eq!(rlp, hex!("f90211a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a01717171717171717171717171717171717171717171717171717171717171717a0171717171717171717171717171717171717171717171717171717171717171780"));